specific class handling of logical operations
